perf(security): Replace 50 sequential includes() with single combined regex in mightContainSecret
Two optimizations to the security keyword pre-filter:

1. Combined regex: Replace 50 sequential String.includes() calls + regex
   test with a single pre-compiled regex alternation. The regex engine
   tests all patterns in one pass through the string with early exit on
   first match, yielding ~3x speedup for the pre-filter function.

2. Module extraction: Move mightContainSecret() and related types to a
   new securityPreFilter.ts module that has no dependency on secretlint.
   Previously, securityCheck.ts imported mightContainSecret from
   securityCheckWorker.ts, which eagerly loaded @secretlint/core (~31ms)
   and @secretlint/secretlint-rule-preset-recommend (~6ms) at module
   import time. Now secretlint is only loaded when actually needed for
   the full lint check.

Benchmark results (50K files, ~3KB each):
- Baseline:  7925ms avg (3 runs)
- Optimized: 6506ms avg (3 runs)
- Improvement: 1419ms (17.9%)

Micro-benchmark (mightContainSecret, 5K files, ~7KB each):
- Old (50x includes): 342ms
- New (combined regex): 114ms
- Speedup: 3.0x (228ms saved)

Module import savings: ~36ms deferred (secretlint loaded on demand)
